backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = "overhead cables and birdlife - backend"
authors = ["Sylvain Sainnier <sainnier@gmail.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
python-decouple = "^3.5"
psycopg2-binary = "^2.9.3"
dj-sinp-nomenclatures = "^0.2.0"
Pillow = "^9.0.0"
djangorestframework-gis = "^0.18"
django-filter = "^21.1"
djoser = "^2.1.0"
coverage = "^6.2"
django-polymorphic = "^3.1.0"
django-rest-polymorphic = "^0.1.9"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
pre-commit = "^2.16.0"


[tool.isort]
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
include = '\.pyi?$'
line-length = 79
target-version = ['py38']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
